requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
click==8.1.2
fastapi==0.75.1
FastAPI-SQLAlchemy==0.2.1
h11==0.13.0
httptools==0.4.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.34
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
